requirements.txt
torchvision
accelerate
diffusers
transformers
numpy
datasets
wandb
black~=23.1
isort>=5.5.4
ruff>=0.0.241,<=0.0.259
tensorflow
tensorflow_datasets==4.6.0
huggingface_hub
Pillow
opencv-python
protobuf==3.20.*